requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2
PyJWT>=2.6
